Folder layout
This commit is intended to layout the folder structure for the Recipe project
on The Ordin Project. The project folder is recipe-project,
then inside it is the index html for where the project starts from, folders images,
scripts and styles added to hold images, javascript files and css files respectively.